requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.32
pydantic==1.9.0
pytest==8.3.3
httpx==0.25.2
requests==2.31.0
pyjwt==2.6.0
passlib[bcrypt]==1.7.4
alembic==1.11.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
websockets==10.0